fullStack/back-end/requirements.txt
sqlalchemy
fastapi
uvicorn[standart]
psycopg2
pydantic
pydantic[email]
passlib
black
bcrypt
starlette
python-jose
alembic
python-dotenv
python-multipart
httpx
pytest
jose
fastapi-cache2[redis]
redis
fastapi-pagination

